requirements.txt
flask
WTForms
mysql-connector-python
flask_wtf
black